pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "shipment-qna-bot"
version = "2.11.24"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-identity>=1.25.1",
    "azure-search-documents>=11.6.0",
    "black>=25.11.0",
    "fastapi>=0.121.3",
    "isort>=7.0.0",
    "langchain>=1.0.8",
    "langchain-core>=1.1.0",
    "langchain-openai>=1.0.3",
    "langgraph>=1.0.3",
    "numpy>=2.3.5",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "pydantic>=2.12.4",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.51.0",
    "tenacity>=9.1.2",
    "uvicorn>=0.38.0",
    "itsdangerous>=2.2.0",
    "azure-storage-blob>=12.28.0",
    "tabulate>=0.9.0",
    "duckdb>=1.1.3",
    "pyarrow>=21.0.0",
    "feedparser>=6.0.12",
    "sgmllib3k>=1.0.0",
]
